Removed TimeInvoker and used the event directly from the scriptable objects
