--- Assets/Scripts/Timer/TimeInvoker.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3cecc8425649c334c8a8d57d631cc976
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
